requirements.txt
Django==4.0.3
django_environ==0.9.0
django_model_helpers==2.0
djangorestframework==3.13.1
environ==1.0
model_helpers==0.3
paypalrestsdk==1.13.1
Pillow==9.2.0
requests==2.27.1
